+++ rust/fury-core/src/resolver/meta_resolver.rs
